requirements.txt
torch
transformers
accelerate
sentencepiece
langchain
ollama
python-dotenv
